Refactored add/remove of favourites and watchlist items to include user account id

+++ app/src/contexts/moviesContext.jsx
@@ -12,13 +12,13 @@ const MoviesContextProvider = (props) => {
   const [watchlist, setWatchlist] = useState(userWatchlist);
   const [myReviews, setMyReviews] = useState( {} )
 
-  const addToFavourites = async (movie) => {
+  const addToFavourites = async (movie, user) => {
     try{
         let updatedFavourites = [...favourites];
         if(!favourites.includes(movie.id)) {
           // add to supabase DB table "favourites"
           let { error } = await supabaseClient
-            .from("favourites").insert({id: movie.id});
+            .from("favourites").insert({id: movie.id, user_id: user});
           // if no error returned from supabase, add to server state
           if(!error){
             updatedFavourites.push(movie.id);
@@ -32,10 +32,10 @@ const MoviesContextProvider = (props) => {
     }
   }
 
-  const removeFromFavourites = async (movie) => {
+  const removeFromFavourites = async (movie, user) => {
     try{
       let { error } = await supabaseClient
-        .from("favourites").delete().eq("id", movie.id);
+        .from("favourites").delete().eq("id", movie.id).eq("user_id", user);
 
       if(!error){
         //remove from server state
@@ -49,12 +49,12 @@ const MoviesContextProvider = (props) => {
     }
   };
 
-  const addToWatchlist = async (movie) => {
+  const addToWatchlist = async (movie, user) => {
     try{
       let updatedWatchlist = [...watchlist];
       if(!watchlist.includes(movie.id)) {
         let { error } = await supabaseClient
-        .from("watchlist").insert({id: movie.id});
+        .from("watchlist").insert({id: movie.id, user_id: user});
       
         if(!error){
           updatedWatchlist.push(movie.id)
@@ -68,10 +68,10 @@ const MoviesContextProvider = (props) => {
     }
   }
 
-  const removeFromWatchlist = async (movie) => {
+  const removeFromWatchlist = async (movie, user) => {
     try{
       let { error } = await supabaseClient
-        .from("watchlist").delete().eq("id", movie.id);
+        .from("watchlist").delete().eq("id", movie.id).eq("user_id", user);
       
       if(!error){
         setWatchlist(watchlist.filter((mId) => mId !== movie.id));
